Backend/requirements.txt
shap
lime>=0.2.0.0
pillow
tensorflow
scikit-learn
scikit-image
Flask
flask-cors
